+++ benchmarks/programs/sieve/main.zig
@@ -1,7 +1,8 @@
 const std = @import("std");
 
 pub fn main() !void {
-    const stdout = std.io.getStdOut().writer();
+    var buf: [4096]u8 = undefined;
+    var stdout = std.fs.File.stdout().writer(&buf);
     var args = std.process.args();
     _ = args.skip();
     const arg = args.next() orelse return;
@@ -22,5 +23,6 @@ pub fn main() !void {
             }
         }
     }
-    try stdout.print("{}\n", .{count});
+    try stdout.interface.print("{}\n", .{count});
+    try stdout.interface.flush();
 }